+++ packages/esm-patient-banner-app/src/hooks/usePatientAttributes.tsx
@@ -1,9 +1,7 @@
-import { useMemo } from 'react';
-import useSWR from 'swr';
-import { FetchResponse, openmrsFetch, useConfig } from '@openmrs/esm-framework';
+import { openmrsFetch, useConfig } from '@openmrs/esm-framework';
 import useSWRImmutable from 'swr/immutable';
 import { ConfigObject } from '../config-schema';
-import { CohortResponse, Patient } from '../types';
+import { Patient } from '../types';
 
 const customRepresentation =
   'custom:(uuid,display,identifiers:(identifier,uuid,preferred,location:(uuid,name),identifierType:(uuid,name,format,formatDescription,validator)),person:(uuid,display,gender,birthdate,dead,age,deathDate,birthdateEstimated,causeOfDeath,preferredName:(uuid,preferred,givenName,middleName,familyName),attributes,preferredAddress:(uuid,preferred,address1,address2,cityVillage,longitude,stateProvince,latitude,country,postalCode,countyDistrict,address3,address4,address5,address6,address7)))';
@@ -44,28 +42,3 @@ export const usePatientContactAttributes = (patientUuid: string) => {
     isLoading,
   };
 };
-
-export function usePatientListsForPatient(patientUuid: string) {
-  const customRepresentation = 'custom:(uuid,patient:ref,cohort:(uuid,name,startDate,endDate))';
-  const url = patientUuid ? `ws/rest/v1/cohortm/cohortmember?patient=${patientUuid}&v=${customRepresentation}` : null;
-  const { data, isLoading, mutate } = useSWR<FetchResponse<CohortResponse>, Error>(url, openmrsFetch);
-
-  const cohorts = useMemo(
-    () =>
-      data
-        ? data?.data?.results.map((ref) => ({
-            uuid: ref.cohort.uuid,
-            name: ref.cohort.name,
-            startDate: ref.cohort.startDate,
-            endDate: ref.cohort.endDate,
-          }))
-        : null,
-    [data],
-  );
-
-  return {
-    data: cohorts,
-    isLoading,
-    mutateLists: mutate,
-  };
-}
